documentation & reducing technical debt
