+++ Washika-dao-frontend/src/components/auth/actions.ts
@@ -1,6 +1,6 @@
 import { useActiveWallet } from "thirdweb/react";
 import { inAppWallet } from "thirdweb/wallets";
-export default function getCurrentConnectedAccount(){
+export default function GetCurrentConnectedAccount(){
     const activeWallet = useActiveWallet();
     if(activeWallet !== null || activeWallet !== undefined){
     return activeWallet;
