+++ Washika-dao-frontend/src/pages/DaoProfile.tsx
@@ -15,7 +15,19 @@ interface DaoDetails {
   multiSigAddr: string;
   kiwango: number;
 }
-
+/**
+ * Component representing the public profile of a DAO (Decentralized Autonomous Organization).
+ * 
+ * This component fetches and displays detailed information about a DAO, including its name,
+ * location, description, and treasury details. It also provides options for DAO operations
+ * such as creating proposals, purchasing shares, requesting loans, and making payments.
+ * 
+ * The component uses the `useParams` hook to retrieve the DAO's multi-signature address
+ * from the URL and fetches the DAO details and member count from the server. It also
+ * adjusts its layout based on the screen size.
+ * 
+ * @returns A React functional component that renders the DAO's public profile page.
+ */
 const PublicDaoProfile: React.FC = () => {
   const navigate = useNavigate();
   const { daoMultiSigAddr } = useParams<{ daoMultiSigAddr: string }>();
