+++ Washika-dao-frontend/src/pages/HomePage.tsx
@@ -16,7 +16,35 @@ interface Blog {
 }
 /**
  * @Auth Policy: Not necessarily authenticated for access but different components require access 
- * @returns 
+ */
+
+/**
+ * Represents the home page component of the application.
+ * 
+ * This component displays various sections including a navigation bar, 
+ * informational content about digital financial groups, and a blog section 
+ * with pagination and search functionality. It allows users to navigate 
+ * to DAO registration and educational pages. Blogs are fetched from a 
+ * JSON file and filtered based on category and search terms.
+ * 
+ * @component
+ * @returns {JSX.Element} The rendered home page component.
+ * 
+ * @remarks
+ * - Utilizes `useState` and `useEffect` hooks for state management and 
+ *   data fetching.
+ * - Implements pagination for blog posts with a default of 3 posts per page.
+ * - Provides search and category filter options for blog posts.
+ * - Includes navigation to DAO registration and educational pages.
+ * 
+ * @interface Blog
+ * @property {string} slug - Unique identifier for the blog.
+ * @property {string} title - Title of the blog.
+ * @property {string} category - Category of the blog.
+ * @property {string} date - Publication date of the blog.
+ * @property {string} [className] - Optional CSS class for styling.
+ * @property {string} [image] - Optional image URL for the blog.
+ * @property {string} [content] - Optional content preview of the blog.
  */
 const HomePage: React.FC = () => {
   const navigate = useNavigate();
